backend/requirements.txt
# Web Framework
fastapi==0.115.0
uvicorn[standard]==0.32.0
python-multipart==0.0.12

# AI/ML
groq==0.9.0
httpx==0.27.0
langchain==0.3.7
langchain-community==0.3.5
chromadb==0.5.18
sentence-transformers==3.2.1

# Document Processing
pypdf2==3.0.1
python-docx==1.1.2
openpyxl==3.1.5
python-pptx==1.0.2
pillow==11.0.0
pytesseract==0.3.13
pdf2image==1.17.0

# Voice
edge-tts==6.1.12

# Utilities
python-dotenv==1.0.1
pydantic==2.9.2
pydantic-settings==2.5.2
numpy==1.26.4

# Testing
pytest==8.3.3
pytest-asyncio==0.24.0
httpx==0.27.0
